+++ src/rank_sel/select_zero_adapt_const.rs
@@ -7,6 +7,7 @@
  */
 
 use super::{Inventory, SpanType};
+use crate::utils::SelectInWord;
 use crate::{
     prelude::{BitCount, BitLength},
     traits::{
@@ -15,7 +16,6 @@ use crate::{
     },
 };
 use ambassador::Delegate;
-use common_traits::SelectInWord;
 use mem_dbg::{MemDbg, MemSize};
 use std::{
     cmp::{max, min},